% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_from_raw.R
\name{index_from_raw}
\alias{index_from_raw}
\title{Calculates index scores from raw domain data input}
\usage{
index_from_raw(
  ds,
  id.col = "id",
  indx = index_table,
  version.col = "ab",
  age.col = "age",
  raw_columns = c("imm", "vis", "ver", "att", "del"),
  mani = TRUE
)
}
\arguments{
\item{ds}{data set}

\item{id.col}{id column name}

\item{indx}{index file}

\item{version.col}{version column name}

\item{age.col}{age column name}

\item{raw_columns}{column names of raw data columns}

\item{mani}{flag to manipulate index data for plotting}
}
\value{
data.frame
}
\description{
Calculates index scores from raw domain data input
}
\examples{
sample_data |> index_from_raw()
}
