% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index.R
\name{plot_index}
\alias{plot_index}
\title{Plot index data. Assumes first column is id}
\usage{
plot_index(
  ds,
  sub_plot = "_is",
  scores = c("_is", "_lo", "_up", "_ci", "_per"),
  dom_names = c("immediate", "visuospatial", "verbal", "attention", "delayed", "total"),
  facet.by = NULL
)
}
\arguments{
\item{ds}{data frame}

\item{sub_plot}{column subset to plot}

\item{scores}{name bits of score variables}

\item{dom_names}{domain names}

\item{facet.by}{variable to base facet_grid on.}
}
\value{
ggplot list object
}
\description{
Plot index data. Assumes first column is id
}
\examples{
ds <- sample_data |> dplyr::filter(ab=="1") |> index_from_raw()
ds |> plot_index(facet.by="ab")
}
