% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_input}
\alias{read_input}
\title{Flexible file import based on extension}
\usage{
read_input(file)
}
\arguments{
\item{file}{file name}
}
\value{
tibble
}
\description{
Flexible file import based on extension
}
