% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index.R
\name{plot_index}
\alias{plot_index}
\title{Plot index data. Assumes first column is id}
\usage{
plot_index(
  data,
  sub_plot = "_is",
  scores = c("_is", "_lo", "_up", "_per"),
  dom_names = c("immediate", "visuospatial", "verbal", "attention", "delayed", "total"),
  facet.by = NULL,
  grp.color = NULL,
  plot.ci = FALSE
)
}
\arguments{
\item{data}{data set}

\item{sub_plot}{column subset to plot. Can be "_is" or "_per".}

\item{scores}{name bits of score variables}

\item{dom_names}{domain names}

\item{facet.by}{variable to base facet_grid on.}

\item{grp.color}{color grouping for ggplot}

\item{plot.ci}{option to include CI in plot}
}
\value{
ggplot list object
}
\description{
Plot index data. Assumes first column is id
}
\examples{
ds <- cognitive.index.lookup::sample_data |> index_from_raw()
ds |>
  tibble::tibble() |>
  plot_index(facet.by = "ab",plot.ci=TRUE)
ds |>
  tibble::tibble() |>
  dplyr::filter(ab == "1")|>
  plot_index(plot.ci=TRUE)
ds |>
  tibble::tibble() |>
  dplyr::filter(ab == "1")|>
  plot_index(sub_plot="_per",plot.ci=TRUE)
data <- ds
}
