% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index.R
\name{plot_index2}
\alias{plot_index2}
\title{Plot index and percentile}
\usage{
plot_index2(ds, ...)
}
\arguments{
\item{ds}{data set}

\item{...}{arguments passed}
}
\value{
ggplot list
}
\description{
Plot index and percentile
}
